Moved to device based system

+++ NesEmu/Core/IDevice.cs
@@ -2,7 +2,7 @@ namespace NesEmu.Core
 {
     public interface IDevice
     {
-        byte Read(byte address);
-        void Write(byte address, byte data);
+        byte Read(ushort address);
+        void Write(ushort address, byte data);
     }
 }
